+++ www/js/src/components/explorers/tables/FieldsTable.js
@@ -1,11 +1,80 @@
 import { useState } from "react";
 
 import "../../../css/Table.css";
+import RadioToggle from "./RadioToggle.jsx";
 
-const FieldsTable = ({ fields, setFields, invalidFields, setInvalidFields }) => {
+const FieldsTable = ({ fields, setFields, invalidFields, setInvalidFields, appModes, setAppModes }) => {
   const [editingCell, setEditingCell] = useState(null);
 
-  const handleFieldTypeChange = (index, newType) => {
+  const modeOptions = [
+    { label: 'Angle', value: 'Angle' },
+    { label: 'Point Source', value: 'PointSource' }
+  ];
+
+  /**
+   * Returns the default values for a field type.
+   * 
+   * @param {string} fieldType - The type of field to get the default values for
+   * @returns {Object} - The default values for the field type
+   */
+  const getFieldTypeDefaultValues = (fieldType) => {
+    switch (fieldType) {
+      case 'Angle':
+        return { Angle: { angle: 0, pupil_sampling: { SquareGrid: { spacing: 0.1 } } } };
+      case 'PointSource':
+        return { PointSource: { x: 0, y: 0, pupil_sampling: { SquareGrid: { spacing: 0.1 } } } };
+      default:
+        console.error(`Unknown field type: ${fieldType}`);
+        return {};
+    }
+  };
+
+  const getSamplingType = (field) => {
+    switch (appModes.fieldType) {
+      case 'Angle':
+        return Object.keys(field.Angle.pupil_sampling)[0];
+      case 'PointSource':
+        return Object.keys(field.PointSource.pupil_sampling)[0];
+      default:
+        console.error(`Unknown mode: ${appModes.fieldType}`);
+        return '';
+    }
+  };
+
+  const handleModeChange = (value) => {
+    const newFields = [...fields];
+
+    switch (value) {
+      case 'Angle':
+        setAppModes(prev => ({ ...prev, fieldType: 'Angle' }));
+
+        // Add default values for Angle fields if they don't exist
+        for (let field of newFields) {
+          if (!field.Angle) {
+            field.Angle = { angle: 0, pupil_sampling: { SquareGrid: { spacing: 0.1 } } };
+          }
+        }
+
+        break;
+      case 'PointSource':
+        setAppModes(prev => ({ ...prev, fieldType: 'PointSource' }));
+
+        // Add default values for PointSource fields if they don't exist
+        for (let field of newFields) {
+          if (!field.PointSource) {
+            field.PointSource = { x: 0, y: 0, pupil_sampling: { SquareGrid: { spacing: 0.1 } } };
+          }
+        }
+        break;
+      default:
+        console.error(`Unknown mode: ${value}`);
+        break;
+    }
+
+    setFields(newFields);
+  };
+
+  const handleSamplingTypeChange = (index, newType) => {
     const newFields = [...fields];
     newFields[index] = { [newType]: { angle: 0, pupil_sampling: { SquareGrid: { spacing: 0.1 } } } };
     setFields(newFields);
@@ -20,12 +89,13 @@ const FieldsTable = ({ fields, setFields, invalidFields, setInvalidFields }) => 
     setEditingCell({ originalValue: value, index, field });
   };
 
-  const handleCellChange = (e, index, field) => {
+  const handleCellChange = (e, index, property) => {
     const newValue = e.target.value;
     const newFields = [...fields];
     const newInvalidFields = { ...invalidFields };
 
-    const invalidStates = (field === "angle" && (newValue < -90.0 || newValue > 90.0))
+    const invalidStates = (property === "angle" && (newValue < -90.0 || newValue > 90.0))
+        || (property === "spacing" && newValue <= 0)
         || isNaN(parseFloat(newValue));
 
     if (invalidStates) {
@@ -33,24 +103,25 @@ const FieldsTable = ({ fields, setFields, invalidFields, setInvalidFields }) => 
         if (!newInvalidFields[index]) {
             newInvalidFields[index] = {};
         }
-        newInvalidFields[index][field] = true;
+        newInvalidFields[index][property] = true;
     } else {
         // A valid number; remove any invalid markers
         if (newInvalidFields[index]) {
-          delete newInvalidFields[index][field];
+          delete newInvalidFields[index][property];
           if (Object.keys(newInvalidFields[index]).length === 0) {
               delete newInvalidFields[index];
           }
         }
     }
 
-    if (field === 'angle') {
+    if (property === 'angle') {
       newFields[index].Angle.angle = newValue;
-    } else if (field === 'spacing') {
+    } else if (property === 'y') {
+      newFields[index].PointSource.y = newValue;
+    } else if (property === 'spacing') {
       newFields[index].Angle.pupil_sampling.SquareGrid.spacing = newValue;
     }
   
-    // TODO: Use reducer hook?
     setFields(newFields);
     setInvalidFields(newInvalidFields);
   };
@@ -77,7 +148,6 @@ const FieldsTable = ({ fields, setFields, invalidFields, setInvalidFields }) => 
       // TODO: Remove the hard-coded "Angle" key as this will be a mess when I add other Field types
       newFields[editingCell.index]["Angle"][editingCell.field] = editingCell.originalValue;
 
-      // TODO: Use reducer hook?
       setFields(newFields);
       setInvalidFields({});
       setEditingCell(null);
@@ -91,7 +161,13 @@ const FieldsTable = ({ fields, setFields, invalidFields, setInvalidFields }) => 
     }
 
     const newFields = [...fields];
-    newFields.splice(index + 1, 0, { Angle: { angle: 0, pupil_sampling: { SquareGrid: { spacing: 0.1 } } } });
+    if (appModes.fieldType === 'Angle') {
+      newFields.splice(index + 1, 0, getFieldTypeDefaultValues('Angle'));
+    }
+    else if (appModes.fieldType === 'PointSource') {
+      newFields.splice(index + 1, 0, getFieldTypeDefaultValues('PointSource'));
+    }
+
     setFields(newFields);
   };
 
@@ -109,28 +185,20 @@ const FieldsTable = ({ fields, setFields, invalidFields, setInvalidFields }) => 
     setFields(newFields);
   };
 
-  const renderFieldTypeCell = (fieldType, index) => {
-        return (
-            <td>
-                <div className="select">  
-                    <select
-                        value={fieldType}
-                        onChange={(e) => handleFieldTypeChange(index, e.target.value)}
-                    >
-                        <option value="Angle">Angle</option>
-                    </select>
-                </div>
-            </td>
-        );
-  };
-
-  const renderSamplingTypeCell = (fieldType, index) => {
+  /**
+   * Renders a cell in the table that allows the user to select the pupil sampling type.
+   * 
+   * @param {string} samplingType - The current pupil sampling type 
+   * @param {number} index - The row index of the cell in the table
+   * @returns {JSX.Element} - The cell element
+   */
+  const renderSamplingTypeCell = (samplingType, index) => {
     return (
         <td>
             <div className="select">  
                 <select
-                    value={fieldType}
-                    onChange={(e) => handleFieldTypeChange(index, e.target.value)}
+                    value={samplingType}
+                    onChange={(e) => handleSamplingTypeChange(index, e.target.value)}
                 >
                     <option value="SquareGrid">Square Grid</option>
                 </select>
@@ -139,9 +207,17 @@ const FieldsTable = ({ fields, setFields, invalidFields, setInvalidFields }) => 
     );
 };
 
-  const renderEditableCell = (value, index, field) => {
-    const isEditing = editingCell && editingCell.index === index && editingCell.field === field;
-    const isInvalid = invalidFields[index] && invalidFields[index][field];
+/**
+ * Renders a cell in the table that can be edited.
+ *
+ * @param {number} value - The value of the cell
+ * @param {number} index - The row index of the cell in the table
+ * @param {string} property - The name of the property in the field object to update
+ * @returns {JSX.Element} - The cell element
+ */
+const renderEditableCell = (value, index, property) => {
+    const isEditing = editingCell && editingCell.index === index && editingCell.field === property;
+    const isInvalid = invalidFields[index] && invalidFields[index][property];
   
     if (isEditing) {
       return (
@@ -150,7 +226,7 @@ const FieldsTable = ({ fields, setFields, invalidFields, setInvalidFields }) => 
             <input
                 type="number"
                 value={value}
-                onChange={(e) => handleCellChange(e, index, field)}
+                onChange={(e) => handleCellChange(e, index, property)}
                 onBlur={handleCellBlur}
                 onKeyDown={handleKeyDown}
                 autoFocus
@@ -160,7 +236,7 @@ const FieldsTable = ({ fields, setFields, invalidFields, setInvalidFields }) => 
     }
     return (
       <div className={`editable-cell ${isInvalid ? 'invalid' : ''}`}>
-        <span onClick={() => handleCellClick(value, index, field)}>
+        <span onClick={() => handleCellClick(value, index, property)}>
           {value}
         </span>
       </div>
@@ -180,30 +256,94 @@ const FieldsTable = ({ fields, setFields, invalidFields, setInvalidFields }) => 
     );
   };
 
-  return (
-    <table className="table is-fullwidth">
-      <thead>
-        <tr>
-          <th className="has-text-weight-semibold has-text-right">Field Type</th>
-          <th className="has-text-weight-semibold has-text-right">Angle</th>
-          <th className="has-text-weight-semibold has-text-right">Pupil Sampling</th>
-          <th className="has-text-weight-semibold has-text-right">Spacing</th>
-          <th>Actions</th>
-        </tr>
-      </thead>
-      <tbody>
-        {fields.map((field, index) => (
-          <tr key={index}>
-            {renderFieldTypeCell(Object.keys(field)[0], index)}
-            <td>{renderEditableCell(field.Angle.angle, index, 'angle')}</td>
-            {renderSamplingTypeCell(Object.keys(field.Angle.pupil_sampling)[0], index)}
-            <td>{renderEditableCell(field.Angle.pupil_sampling.SquareGrid.spacing, index, 'spacing')}</td>
-            {renderActionButtons(index)}
-          </tr>
-        ))}
-      </tbody>
-    </table>
-  );
+  const renderComponentDependingOnActiveFieldType = () => {
+    const fieldType = appModes.fieldType;
+    const fieldTypeHeader = appModes.fieldType === 'Angle' ? 'Angle' : 'Point Source';
+
+    const data = (field) => {
+      switch (fieldType) {
+        case 'Angle':
+          return field.Angle.angle
+        case 'PointSource':
+          return field.PointSource.y
+        default:
+          console.error(`Unknown field type: ${fieldType}`);
+          return {};
+      }
+    }
+
+    /**
+     * Determines the pupil sampling type for a field.
+     * @param {Object} field - The field object 
+     * @returns {string} - The pupil sampling type
+     */
+    const pupilSampling = (field) => {
+      switch (fieldType) {
+        case 'Angle':
+          return Object.keys(field.Angle.pupil_sampling)[0]
+        case 'PointSource':
+          return Object.keys(field.PointSource.pupil_sampling)[0]
+        default:
+          console.error(`Unknown field type: ${fieldType}`);
+          return {};
+      }
+    }
+
+    const spacing = (field) => {
+      switch (fieldType) {
+        case 'Angle':
+          return field.Angle.pupil_sampling.SquareGrid.spacing
+        case 'PointSource':
+          return field.PointSource.pupil_sampling.SquareGrid.spacing
+        default:
+          console.error(`Unknown field type: ${fieldType}`);
+          return {};
+      }
+    }
+    const propertyName = fieldType === 'Angle' ? 'angle' : 'y';
+
+    return (
+      <div className="fields-table">
+  
+        <div className="has-background-light py-2">
+            <div className="container">
+                <div className="is-flex is-justify-content-center">
+                    <RadioToggle
+                        options={modeOptions}
+                        selectedValue={appModes.fieldType}
+                        onChange={handleModeChange}
+                        name="fieldType"
+                        className="is-flex-direction-row"
+                    />
+              </div>
+            </div>
+        </div>
+  
+        <table className="table is-fullwidth">
+          <thead>
+            <tr>
+              <th className="has-text-weight-semibold has-text-right">{fieldTypeHeader}</th>
+              <th className="has-text-weight-semibold has-text-right">Pupil Sampling</th>
+              <th className="has-text-weight-semibold has-text-right">Spacing</th>
+              <th>Actions</th>
+            </tr>
+          </thead>
+          <tbody>
+            {fields.map((field, index) => (
+              <tr key={index}>
+                <td>{renderEditableCell(data(field), index, propertyName)}</td>
+                {renderSamplingTypeCell(pupilSampling(field), index)}
+                <td>{renderEditableCell(spacing(field), index, 'spacing')}</td>
+                {renderActionButtons(index)}
+              </tr>
+            ))}
+          </tbody>
+        </table>
+      </div>
+    );
+  };
+
+  return renderComponentDependingOnActiveFieldType();
 };
 
 export default FieldsTable;