+++ www/js/src/examples/convexplanoLens.js
@@ -14,7 +14,7 @@ const aperture = {"EntrancePupil": { "semi_diameter": 5.0 }};
 
 const wavelengths = [0.5876];
 
-const appModes = { "refractiveIndex": true };
+const appModes = { fieldType: "Angle", "refractiveIndex": true };
 
 const exampleData = {
     "surfaces": surfaces,